- removed SPI classes (now in gloop-spi and implemented in gloop-impl-opengl) - fixed return type for getTextureMaxAnisotropy

+++ src/main/java/com/longlinkislong/gloop/GLTextureParameters.java
@@ -389,7 +389,7 @@ public class GLTextureParameters {
                 return this.maxLevel;
             }
 
-            this.maxLevel = GLTools.getDriverInstance().textureGetMaxAnisotropy();            
+            this.maxLevel = (float) GLTools.getDriverInstance().textureGetMaxAnisotropy();            
             this.checked = true;
 
             LOGGER.trace(
